Relocate the helper fucntion.
